Added user authentication

+++ app/api/ratings/[id]/route.js
@@ -1,9 +1,19 @@
 // http://localhost:3000/api/ratings/1234
 import prisma from "@/app/libs/prismadb";
+import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export const GET = async (req, { params }) => {
   try {
+    //Check if user is auth
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return new NextResponse(JSON.stringify({ error: "unauthorized" }), {
+        status: 401,
+      });
+    }
+
     const { id } = params;
     const rating = await prisma.rating.findUnique({
       where: {
@@ -26,6 +36,14 @@ export const GET = async (req, { params }) => {
 
 export const POST = async (req, { params }) => {
   try {
+    //Check if user is auth
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return new NextResponse(JSON.stringify({ error: "unauthorized" }), {
+        status: 401,
+      });
+    }
+
     const body = await req.json();
     const { id } = params;
     const { title, scary, story, acting } = body;
@@ -56,6 +74,14 @@ export const POST = async (req, { params }) => {
 
 export const PATCH = async (req, { params }) => {
   try {
+    //Check if user is auth
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return new NextResponse(JSON.stringify({ error: "unauthorized" }), {
+        status: 401,
+      });
+    }
+
     const body = await req.json();
     const { id } = params;
     const { title, scary, story, acting } = body;
@@ -86,6 +112,14 @@ export const PATCH = async (req, { params }) => {
 
 export const DELETE = async (req, { params }) => {
   try {
+    //Check if user is auth
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return new NextResponse(JSON.stringify({ error: "unauthorized" }), {
+        status: 401,
+      });
+    }
+
     const { id } = params;
     const rating = await prisma.rating.delete({
       where: {
